Level Select
Level Select Works and will disable selecting a level you have not yet unlocked. Working on making a pause menu next. Had to make some new scripts that are just for the main menu. Trying to work on player prefs but may not work so we dont know just yet!